+++ kleidicv/src/filters/median_blur_sc.h
@@ -6,17 +6,17 @@
 #define KLEIDICV_MEDIAN_BLUR_SC_H
 
 #include <algorithm>
-#include <utility>
 
 #include "kleidicv/ctypes.h"
-#include "kleidicv/filters/filter_2d.h"
-#include "kleidicv/filters/filter_2d_5x5_sc.h"
-#include "kleidicv/filters/filter_2d_7x7_sc.h"
+#include "kleidicv/filters/filter_2d_sc.h"
 #include "kleidicv/filters/median_blur.h"
+#include "kleidicv/filters/process_fitler_2d.h"
 #include "kleidicv/kleidicv.h"
 #include "kleidicv/sve2.h"
+#include "kleidicv/workspace/border_3x3.h"
 #include "kleidicv/workspace/border_5x5.h"
 #include "kleidicv/workspace/border_7x7.h"
+#include "median_blur_sorting_network_3x3.h"
 #include "median_blur_sorting_network_5x5.h"
 #include "median_blur_sorting_network_7x7.h"
 namespace KLEIDICV_TARGET_NAMESPACE {
@@ -26,36 +26,69 @@ template <typename ScalarType, size_t KernelSize>
 class MedianBlur;
 
 template <typename ScalarType>
-class MedianBlurBase {
- protected:
-  class vectorized_comparator {
-   public:
-    using SourceVectorType = typename VecTraits<ScalarType>::VectorType;
+class VectorComparator {
+ public:
+  using SourceVectorType = typename VecTraits<ScalarType>::VectorType;
 
-    static void compare_and_swap(SourceVectorType& left,
-                                 SourceVectorType& right,
-                                 svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
-      SourceVectorType max_value = svmax_m(pg, left, right);
-      SourceVectorType min_value = svmin_m(pg, left, right);
-      left = min_value;
-      right = max_value;
-    }
+  static void compare_and_swap(SourceVectorType& left, SourceVectorType& right,
+                               svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
+    SourceVectorType max_value = svmax_m(pg, left, right);
+    SourceVectorType min_value = svmin_m(pg, left, right);
+    left = min_value;
+    right = max_value;
+  }
 
-    static void min(SourceVectorType& left, SourceVectorType& right,
-                    svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
-      left = svmin_m(pg, left, right);
-    }
+  static void min(SourceVectorType& left, SourceVectorType& right,
+                  svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
+    left = svmin_m(pg, left, right);
+  }
 
-    static void max(SourceVectorType& left, SourceVectorType& right,
-                    svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
-      right = svmax_m(pg, left, right);
-    }
-  };
+  static void max(SourceVectorType& left, SourceVectorType& right,
+                  svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
+    right = svmax_m(pg, left, right);
+  }
+  static SourceVectorType get_min(SourceVectorType& left,
+                                  SourceVectorType& right,
+                                  svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
+    return svmin_m(pg, left, right);
+  }
+  static SourceVectorType get_max(SourceVectorType& left,
+                                  SourceVectorType& right,
+                                  svbool_t& pg) KLEIDICV_STREAMING_COMPATIBLE {
+    return svmax_m(pg, left, right);
+  }
 };
+
+// Template for Median Blur 3x3 filters.
+template <typename ScalarType>
+class MedianBlur<ScalarType, 3> {
+ public:
+  using SourceType = ScalarType;
+  using DestinationType = SourceType;
+  using SourceVectorType = typename VecTraits<SourceType>::VectorType;
+  using DestinationVectorType = typename VecTraits<DestinationType>::VectorType;
+
+  template <typename KernelWindowFunctor>
+  void vector_path(svbool_t& pg, KernelWindowFunctor& KernelWindow,
+                   DestinationVectorType& output_vec) const
+      KLEIDICV_STREAMING_COMPATIBLE {
+    sorting_network3x3_single_row<VectorComparator<ScalarType>>(KernelWindow,
+                                                                output_vec, pg);
+  }
+
+  template <typename KernelWindowFunctor>
+  void vector_path_for_dual_row_handling(
+      svbool_t& pg, KernelWindowFunctor& KernelWindow,
+      DestinationVectorType& output_vec_0,
+      DestinationVectorType& output_vec_1) const KLEIDICV_STREAMING_COMPATIBLE {
+    sorting_network3x3_dual_rows<VectorComparator<ScalarType>>(
+        KernelWindow, output_vec_0, output_vec_1, pg);
+  }
+};  // end of class MedianBlur<ScalarType, 3>
 
 // Template for Median Blur 5x5 filters.
 template <typename ScalarType>
-class MedianBlur<ScalarType, 5> : public MedianBlurBase<ScalarType> {
+class MedianBlur<ScalarType, 5> {
  public:
   using SourceType = ScalarType;
   using DestinationType = SourceType;
@@ -64,19 +97,19 @@ class MedianBlur<ScalarType, 5> : public MedianBlurBase<ScalarType> {
   using SourceVectorType = typename SourceVecTraits::VectorType;
   using DestinationVectorType = typename KLEIDICV_TARGET_NAMESPACE::VecTraits<
       DestinationType>::VectorType;
-  using VectorComparator =
-      typename MedianBlurBase<SourceType>::vectorized_comparator;
+
   template <typename KernelWindowFunctor>
-  void vector_path(KernelWindowFunctor& KernelWindow,
-                   DestinationVectorType& output_vec,
-                   svbool_t& pg) const KLEIDICV_STREAMING_COMPATIBLE {
-    sorting_network5x5<VectorComparator>(KernelWindow, output_vec, pg);
+  void vector_path(svbool_t& pg, KernelWindowFunctor& KernelWindow,
+                   DestinationVectorType& output_vec) const
+      KLEIDICV_STREAMING_COMPATIBLE {
+    sorting_network5x5<VectorComparator<ScalarType>>(KernelWindow, output_vec,
+                                                     pg);
   }
 };  // end of class MedianBlur<ScalarType, 5>
 
 // Template for Median Blur 7x7 filters.
 template <typename ScalarType>
-class MedianBlur<ScalarType, 7> : public MedianBlurBase<ScalarType> {
+class MedianBlur<ScalarType, 7> {
  public:
   using SourceType = ScalarType;
   using DestinationType = SourceType;
@@ -85,14 +118,13 @@ class MedianBlur<ScalarType, 7> : public MedianBlurBase<ScalarType> {
   using SourceVectorType = typename SourceVecTraits::VectorType;
   using DestinationVectorType = typename KLEIDICV_TARGET_NAMESPACE::VecTraits<
       DestinationType>::VectorType;
-  using VectorComparator =
-      typename MedianBlurBase<SourceType>::vectorized_comparator;
 
   template <typename KernelWindowFunctor>
-  void vector_path(KernelWindowFunctor& KernelWindow,
-                   DestinationVectorType& output_vec,
-                   svbool_t& pg) const KLEIDICV_STREAMING_COMPATIBLE {
-    sorting_network7x7<VectorComparator>(KernelWindow, output_vec, pg);
+  void vector_path(svbool_t& pg, KernelWindowFunctor& KernelWindow,
+                   DestinationVectorType& output_vec) const
+      KLEIDICV_STREAMING_COMPATIBLE {
+    sorting_network7x7<VectorComparator<ScalarType>>(KernelWindow, output_vec,
+                                                     pg);
   }
 };  // end of class MedianBlur<ScalarType, 7>
 
@@ -105,18 +137,24 @@ kleidicv_error_t median_blur_stripe_sc(
   Rectangle rect{width, height};
   Rows<const T> src_rows{src, src_stride, channels};
   Rows<T> dst_rows{dst, dst_stride, channels};
-  if (kernel_width == 5) {
+  if (kernel_width == 3) {
+    MedianBlur<T, 3> median_filter;
+    Filter2D3x3<MedianBlur<T, 3>> filter{median_filter};
+    process_filter2d_by_dual_rows(rect, y_begin, y_end, src_rows, dst_rows,
+                                  border_type, filter);
+  } else if (kernel_width == 5) {
     MedianBlur<T, 5> median_filter;
     Filter2D5x5<MedianBlur<T, 5>> filter{median_filter};
     process_filter2d(rect, y_begin, y_end, src_rows, dst_rows, border_type,
                      filter);
     return KLEIDICV_OK;
+  } else {
+    MedianBlur<T, 7> median_filter;
+    Filter2D7x7<MedianBlur<T, 7>> filter{median_filter};
+    process_filter2d(rect, y_begin, y_end, src_rows, dst_rows, border_type,
+                     filter);
   }
 
-  MedianBlur<T, 7> median_filter;
-  Filter2D7x7<MedianBlur<T, 7>> filter{median_filter};
-  process_filter2d(rect, y_begin, y_end, src_rows, dst_rows, border_type,
-                   filter);
   return KLEIDICV_OK;
 }